+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Head
 
+- Added: `selector-pseudo-class-no-unknown` rule.
 - Fixed: `declaration-block-no-ignored-properties` now detects use of `min-width` and `max-width` with inline, table-row, table-row-group, table-column and table-column-group elements.
 
 # 6.3.3